Create a new package to handle CSRF tokens
Added a new handler for CSRF tokens, that
is easy to include into a "vanilla" or at least
a non-framework project.
Added a gitignore file, with "vendor" as entry.
Updated README for how to use package.
Added a bunch of tests, to check if it actually
works as intended :)

Based on finding found here:
https://brightsec.com/blog/csrf-token/