We now have a generator and a token object
Every token generated is now stored in a unique
session key. This is done to hopefully improve
on security, as external scripts cannot mimic or scrape
the forms that easially. (they can - but its
harder).

Updated README.md with new code examples.
Updated unittest.
Renamed the Csrf class to Generator.
Added a Token class to handle a value and a key.
Removed composer.lock, since it should not be
added to git :)